+++ RELEASE_NOTES.md
@@ -1,3 +1,15 @@
+# v0.9.1
+xxxx-yy-zz
+* Several improvements to error types:
+  * Removed the deprecated `description` function from `Error` impl for error types.
+  * Added a `cause()` method to the Error impl for error types if they have any variants which
+    contain error values.
+  * Improved the `Display` impl for Error types by using the documentation strings provided,
+    including recursing into nested errors or values if present.
+  * Routes with no error type defined now return a new `NoError` type instead of `()`. This new type
+    implements `Error`, so now all routes' error types are guaranteed to be `Error`, and hence,
+    `Display` as well. (You should never actually see instances of this type.)
+
 # v0.9.0
 2020-11-30
 * All structs and enums except ones marked `union_closed` in the spec are now marked with